add opening files, play 100 matches of 1 sec add a check to be able to run the script locally

+++ test/EloRegressionTest/eloRegressionTest.sh
@@ -2,31 +2,36 @@
 
 echo "Starting ELO testing script"
 
-#Copying current version to the test folder
+##Copying current version to the test folder
 cp ../../Deepov DeepovCurrent
 
-git clone --branch=master git://github.com/RomainGoussault/Deepov.git LatestRelease/
-cd LatestRelease/
-mkdir obj
+if [ ! -f Deepov ]; then
 
-# Get new tags from remote
-git fetch --tags 
+    echo "Deepov not found. Cloning from github and building"
 
-# Get latest tag name
-latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
+	git clone --branch=master git://github.com/RomainGoussault/Deepov.git LatestRelease/
+	cd LatestRelease/
+	mkdir obj
 
-# Checkout latest tag
-git checkout $latestTag
+	# Get new tags from remote
+	git fetch --tags 
 
-#Compile it and make it executable
-make Deepov
-chmod +x Deepov
+	# Get latest tag name
+	latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
 
-#Copying latest release version to the test folder
-cp Deepov ../Deepov
-cd ..
+	# Checkout latest tag
+	git checkout $latestTag
 
-./cutechess-cli.sh -engine cmd=DeepovCurrent -engine cmd=Deepov -each proto=uci tc=100/10 -rounds 10 -pgnout matchOuput.txt -recover -repeat | while read line; do
+	#Compile it and make it executable
+	make Deepov
+	chmod +x Deepov
+
+	#Copying latest release version to the test folder
+	cp Deepov ../Deepov
+	cd ..
+fi
+
+./cutechess-cli.sh -engine cmd=DeepovCurrent -engine cmd=Deepov -each proto=uci tc=100/1 -rounds 100 -pgnout matchOuput.txt -recover -repeat -openings file=openings.pgn | while read line; do
 		
 	echo "$line"
